perf(ci): split heavy test files for balanced shard distribution
