+++ .github/workflows/node-test.yml
@@ -1,7 +1,7 @@
-# CI Speed Optimisation U1: Shard node tests into 10 parallel matrix jobs.
+# CI Speed Optimisation: Shard node tests into 15 parallel matrix jobs.
 #
-# The 626-file Node test suite is split across 10 shards using
-# `node --test --test-shard=N/10` (built-in since Node 22). Each shard
+# The 630+ file Node test suite is split across 15 shards using
+# `node --test --test-shard=N/15` (built-in since Node 22). Each shard
 # runs in its own GitHub Actions matrix cell, reducing wall-clock from
 # ~700s (single job) to < 60s.
 #
@@ -62,7 +62,7 @@ jobs:
     needs: changes
     if: ${{ github.event_name == 'workflow_dispatch' || needs.changes.outputs.docs_only != 'true' }}
     runs-on: ubuntu-latest
-    timeout-minutes: 7
+    timeout-minutes: 5
     strategy:
       fail-fast: false
       matrix:
