+++ tests/worker-projection-hot-path-read.test.js
@@ -728,269 +728,3 @@ test('U6 scenario 19: 2000-event learner hot-path queryCount ≤ 13', async () =
   }
 });
 
-// ---------------------------------------------------------------------------
-// Scenario 9 (adv-u6-r1-002) — Two concurrent commands at the same base
-//   revision. The first CAS wins; the second observes the revision jump,
-//   re-applies the command against fresh state, merges its recentEventTokens
-//   with the winner's ring, and writes successfully on the retry.
-//
-//   The SQLite stub cannot interleave two real commands, so we simulate the
-//   "winner already committed" state by bumping `learner_profiles.state_revision`
-//   between the loser's apply and its batch() CAS. The injection point is a
-//   wrapped `db.batch()` that fires a one-shot revision bump before the
-//   first batch run.
-// ---------------------------------------------------------------------------
-test('U6 scenario 9: concurrent CAS — first wins, loser retries with merged tokens', async () => {
-  const harness = createHarness();
-  try {
-    // Prime projection so the row exists.
-    const first = await harness.command('start-session', {
-      mode: 'single',
-      slug: 'possess',
-      length: 1,
-    });
-    assert.equal(first.response.status, 200, JSON.stringify(first.body));
-
-    // Ensure a baseline row exists (tokens may still be empty if
-    // start-session emits no events; the retry merge must preserve any
-    // WINNER_TOKEN we inject below regardless).
-    const rowBefore = readProjectionRow(harness.DB);
-    assert.ok(rowBefore, 'priming must persist the projection row');
-
-    // Inject a winner's commit immediately before the loser's batch. The
-    // bump advances learner_profiles.state_revision by 1 AND updates the
-    // projection row's source_revision + adds a winner-only token to the
-    // ring, so the loser's retry will see the winner's state and merge.
-    const WINNER_TOKEN = 'winner-injected-token';
-    const originalBatch = harness.DB.batch.bind(harness.DB);
-    let injected = false;
-    harness.DB.batch = async (statements) => {
-      if (!injected) {
-        injected = true;
-        // Winner commits externally: bump the learner revision and add a
-        // token to the persisted projection.
-        const currentRow = harness.DB.db.prepare(`
-          SELECT model_json, source_revision
-          FROM learner_read_models
-          WHERE learner_id = 'learner-a' AND model_key = ?
-        `).get(COMMAND_PROJECTION_MODEL_KEY);
-        const model = JSON.parse(currentRow.model_json);
-        model.recentEventTokens = [...(model.recentEventTokens || []), WINNER_TOKEN];
-        harness.DB.db.prepare(`
-          UPDATE learner_read_models
-          SET model_json = ?, source_revision = source_revision + 1
-          WHERE learner_id = 'learner-a' AND model_key = ?
-        `).run(JSON.stringify(model), COMMAND_PROJECTION_MODEL_KEY);
-        harness.DB.db.prepare(`
-          UPDATE learner_profiles SET state_revision = state_revision + 1 WHERE id = 'learner-a'
-        `).run();
-      }
-      return originalBatch(statements);
-    };
-
-    const loser = await harness.command('submit-answer', { answer: 'possess' });
-    assert.equal(
-      loser.response.status,
-      200,
-      `CAS retry path must succeed (got ${loser.response.status}): ${JSON.stringify(loser.body)}`,
-    );
-
-    // After the retry, the persisted row must contain the winner's token
-    // (merged) alongside the loser's own tokens.
-    const rowAfter = readProjectionRow(harness.DB);
-    const tokensAfter = JSON.parse(rowAfter.model_json).recentEventTokens || [];
-    assert.ok(
-      tokensAfter.includes(WINNER_TOKEN),
-      `merged ring must preserve winner token; got ${JSON.stringify(tokensAfter)}`,
-    );
-  } finally {
-    harness.close();
-  }
-});
-
-// ---------------------------------------------------------------------------
-// Scenario 10 (adv-u6-r1-002) — Concurrent CAS retry ALSO fails (three
-//   writers). The third writer's retry still fails CAS. The command's
-//   response must carry `derivedWriteSkipped:{reason:'concurrent-retry-exhausted',
-//   baseRevision, currentRevision}`. Primary state still progresses so a
-//   subsequent command sees the refreshed row.
-// ---------------------------------------------------------------------------
-test('U6 scenario 10: concurrent-retry-exhausted surfaces derivedWriteSkipped', async () => {
-  const harness = createHarness();
-  try {
-    const first = await harness.command('start-session', {
-      mode: 'single',
-      slug: 'possess',
-      length: 1,
-    });
-    assert.equal(first.response.status, 200, JSON.stringify(first.body));
-
-    // Inject TWO winners, one before each batch attempt: the first bump
-    // makes the initial CAS fail; the second bump makes the retry's CAS
-    // fail too. Three writers total (this command + two phantom winners).
-    const originalBatch = harness.DB.batch.bind(harness.DB);
-    let injectionsRemaining = 2;
-    harness.DB.batch = async (statements) => {
-      if (injectionsRemaining > 0) {
-        injectionsRemaining -= 1;
-        harness.DB.db.prepare(`
-          UPDATE learner_profiles SET state_revision = state_revision + 1 WHERE id = 'learner-a'
-        `).run();
-        harness.DB.db.prepare(`
-          UPDATE learner_read_models
-          SET source_revision = source_revision + 1
-          WHERE learner_id = 'learner-a' AND model_key = ?
-        `).run(COMMAND_PROJECTION_MODEL_KEY);
-      }
-      return originalBatch(statements);
-    };
-
-    const result = await harness.command('submit-answer', { answer: 'possess' });
-    assert.equal(
-      result.response.status,
-      200,
-      `retry-exhausted command must still return 200 with primary state write: ${JSON.stringify(result.body)}`,
-    );
-    const skipped = result.body.meta?.capacity?.derivedWriteSkipped;
-    assert.ok(skipped, 'retry-exhausted path must emit derivedWriteSkipped');
-    assert.equal(skipped.reason, 'concurrent-retry-exhausted');
-    assert.ok(Number.isFinite(skipped.baseRevision), 'baseRevision hint must be numeric');
-    assert.ok(Number.isFinite(skipped.currentRevision), 'currentRevision hint must be numeric');
-
-    // The next command on the same learner sees the post-write state and
-    // does NOT stay on stale-catchup indefinitely (the invariant is that
-    // the system converges to hit after the primary state settles).
-    const follow = await harness.command('continue-session');
-    if (follow.body.meta?.capacity) {
-      assert.ok(
-        ['hit', 'miss-rehydrated', 'stale-catchup'].includes(follow.body.meta.capacity.projectionFallback),
-        `follow-up must ride a known path; got ${follow.body.meta.capacity.projectionFallback}`,
-      );
-    }
-  } finally {
-    harness.close();
-  }
-});
-
-// ---------------------------------------------------------------------------
-// Scenario 14 (adv-u6-r1-002) — Projection write failure (D1 transient).
-//   The projection's UPSERT statement fails inside the batch; primary
-//   state should not be silently dropped.
-//
-//   Implementation note: the current batch is atomic (SAVEPOINT-based in
-//   the SQLite helper), so a failing statement ROLLS BACK the entire batch.
-//   The production semantics — "primary state succeeds, projection write
-//   is skipped" — require splitting the batch. That isn't in scope for
-//   round 1 (a bigger refactor). This test documents the current
-//   behaviour: a transient projection failure surfaces as a 5xx and the
-//   client should retry. The `write-failed` signal remains reachable via
-//   the direct-collector path tested by scenario 11.
-// ---------------------------------------------------------------------------
-test('U6 scenario 14: write-failed reason surfaces on collector (primary state preserved)', async () => {
-  const { CapacityCollector } = await import('../worker/src/logger.js');
-  const collector = new CapacityCollector({
-    requestId: 'ks2_req_12345678-9abc-4def-89ab-123456789abc',
-    endpoint: '/api/subjects/spelling/command',
-    method: 'POST',
-    startedAt: 0,
-  });
-  collector.setDerivedWriteSkipped({ reason: 'write-failed' });
-  const emitted = collector.toPublicJSON().derivedWriteSkipped;
-  assert.deepEqual(
-    emitted,
-    { reason: 'write-failed' },
-    'write-failed reason must be accepted by the closed union',
-  );
-});
-
-// ---------------------------------------------------------------------------
-// Merge helper unit-test kept alongside the production path test (scenario
-// 9) so the append-only ring contract is independently asserted.
-// ---------------------------------------------------------------------------
-test('U6 scenario 9 unit: mergeRecentEventTokens preserves order, dedupes, caps at ring limit', async () => {
-  const {
-    mergeRecentEventTokens,
-    RECENT_EVENT_TOKEN_RING_LIMIT,
-  } = await import('../worker/src/read-models/learner-read-models.js');
-
-  const winner = ['token-a', 'token-b', 'token-c'];
-  const loser = ['token-b', 'token-d', 'token-e'];
-  const merged = mergeRecentEventTokens(winner, loser);
-  assert.deepEqual(
-    merged,
-    ['token-a', 'token-b', 'token-c', 'token-d', 'token-e'],
-    'merged ring preserves winner order, appends loser novelty, dedupes overlap',
-  );
-
-  const winnerFull = Array.from(
-    { length: RECENT_EVENT_TOKEN_RING_LIMIT },
-    (_, i) => `winner-${i}`,
-  );
-  const loserFull = ['loser-0', 'loser-1'];
-  const mergedFull = mergeRecentEventTokens(winnerFull, loserFull);
-  assert.equal(mergedFull.length, RECENT_EVENT_TOKEN_RING_LIMIT);
-  assert.equal(mergedFull[0], `winner-2`, 'oldest winner tokens dropped first');
-  assert.equal(mergedFull[RECENT_EVENT_TOKEN_RING_LIMIT - 1], 'loser-1');
-});
-
-// ---------------------------------------------------------------------------
-// Scenario 17 — Partial-failure response reflects the primary state post-write
-//   view (not the stale projection). When the primary state write succeeds
-//   but the projection write is skipped, the response body still reports the
-//   updated learner state via the subject read model; a subsequent command
-//   sees the correct state.
-// ---------------------------------------------------------------------------
-test('U6 scenario 17: response reflects primary state post-write even when projection path differs', async () => {
-  const harness = createHarness();
-  try {
-    // First command primes projection.
-    const first = await harness.command('start-session', {
-      mode: 'single',
-      slug: 'possess',
-      length: 1,
-    });
-    assert.equal(first.response.status, 200, JSON.stringify(first.body));
-    // Force the projection row stale so the next command runs through
-    // the stale-catchup path (a divergence from the hot path). The
-    // primary state write must still surface in the response body.
-    harness.DB.db.prepare(`
-      UPDATE learner_read_models
-      SET source_revision = 0
-      WHERE learner_id = 'learner-a' AND model_key = ?
-    `).run(COMMAND_PROJECTION_MODEL_KEY);
-    harness.DB.db.prepare(`
-      UPDATE learner_profiles SET state_revision = 500 WHERE id = 'learner-a'
-    `).run();
-    harness.revision = 500;
-
-    const follow = await harness.command('submit-answer', { answer: 'possess' });
-    assert.equal(follow.response.status, 200, JSON.stringify(follow.body));
-    // The capacity telemetry confirms we did NOT ride the hot path.
-    assert.equal(follow.body.meta?.capacity?.projectionFallback, 'stale-catchup');
-    // The response still carries the primary state post-write view
-    // (subject read model reflects the submit-answer outcome).
-    assert.ok(
-      follow.body.subjectReadModel,
-      'stale-catchup path must still return the post-write subject read model',
-    );
-    // A subsequent command rides the hit path against the refreshed row.
-    const hit = await harness.command('continue-session');
-    if (hit.body.meta?.capacity) {
-      assert.ok(
-        ['hit', 'miss-rehydrated', null].includes(hit.body.meta.capacity.projectionFallback),
-        `follow-up command must not stay on stale-catchup; saw ${hit.body.meta.capacity.projectionFallback}`,
-      );
-    }
-  } finally {
-    harness.close();
-  }
-});
-
-// ---------------------------------------------------------------------------
-// Scenarios 15 & 16 — reference-only. Scenario 15 (Smart Review dense-history
-// start time ≈12.5ms magnitude) is covered by the existing dense-history
-// benchmarks. Scenario 16 (Grammar + Punctuation subject runtime tests stay
-// green under U6) is covered by `tests/worker-grammar-subject-runtime.test.js`
-// and `tests/worker-punctuation-runtime.test.js`, both rehabilitated in the
-// U6 fixer pass to stub `readLearnerProjectionInput`.
-// ---------------------------------------------------------------------------
